refactor: split runtime.go
